// packages/source-sync/src/lambda-modules/sync-boundary/infrastructure/adapter/out/db/typeorm/fields/mt-field-assignment-columns.enum.ts
export enum MtFieldAssignmentColumns {
  ID = 'ID',
  DOMAIN_ID = 'DOMAIN_ID',
  WORK_ORDER_ID = 'WORK_ORDER_ID',
  COMPANY = 'COMPANY',
  TECHNICIAN = 'TECHNICIAN',
  STARTED_AT = 'STARTED_AT',
  ENDED_AT = 'ENDED_AT',
  MANUALLY_ADDED = 'MANUALLY_ADDED',
  CREATED_AT = 'CREATED_AT',
  UPDATED_AT = 'UPDATED_AT',
}